Allow game code to override BM Font parsing and support rotated font atlas textures (#20300)
* Added RenderTexture::saveToFileAsNonPMA() to save images without PMA.
Set the PMA parameter to true when calling initWithRawData() inside RenderTexture::newImage(), since textures are PMA.
Renamed Image::premultipliedAlpha() to Image::premultiplyAlpha() to better reflect it's action, and made it public.
Added Image::reversePremultipliedAlpha() to allow the reversing of the PMA.
Updated CCImage-ios.mm to set the correct bitmapInfo for PMA and non-PMA images before saving a file.
Updated RenderTextureTest::RenderTextureSave() to cater for non-PMA file saving.

* [CCImage-ios.mm] Fixed indentation.

* [CCFontFNT.cpp/.h] Allow game to override FNT file parsing for extended formats.
[CCFontAtlas] Added flag to indicated if a texture is rotated in FontLetterDefinition.  This will allow for embedding a font atlas image inside a larger image atlas that supports rotation.
[CCLabel.cpp] Ensure that the rotated flag in the font letter definition is used when creating a sprite for the letter.

* [CCFontFNT.h] Fixed missing include statements.
