+++ cocos/2d/CCFontAtlas.cpp
@@ -436,6 +436,7 @@ bool FontAtlas::prepareLetterDefinitions(const std::u32string& utf32Text)
             tempDef.height = tempDef.height / scaleFactor;
             tempDef.U = tempDef.U / scaleFactor;
             tempDef.V = tempDef.V / scaleFactor;
+            tempDef.rotated = false;
         }
         else{
             delete[] bitmap;
@@ -451,6 +452,7 @@ bool FontAtlas::prepareLetterDefinitions(const std::u32string& utf32Text)
             tempDef.offsetX = 0;
             tempDef.offsetY = 0;
             tempDef.textureID = 0;
+            tempDef.rotated = false;
             _currentPageOrigX += 1;
         }
