+++ cocos/2d/CCFontFNT.h
@@ -30,10 +30,106 @@
 /// @cond DO_NOT_SHOW
 
 #include "2d/CCFont.h"
+#include <set>
+#include <unordered_map>
 
 NS_CC_BEGIN
 
-class BMFontConfiguration;
+/**
+@struct BMFontDef
+BMFont definition
+*/
+typedef struct _BMFontDef {
+    //! ID of the character
+    unsigned int charID;
+    //! origin and size of the font
+    Rect rect;
+    //! The X amount the image should be offset when drawing the image (in pixels)
+    short xOffset;
+    //! The Y amount the image should be offset when drawing the image (in pixels)
+    short yOffset;
+    //! The amount to move the current position after drawing the character (in pixels)
+    short xAdvance;
+} BMFontDef;
+
+/** @struct BMFontPadding
+BMFont padding
+@since v0.8.2
+*/
+typedef struct _BMFontPadding {
+    /// padding left
+    int left;
+    /// padding top
+    int top;
+    /// padding right
+    int right;
+    /// padding bottom
+    int bottom;
+} BMFontPadding;
+
+/** @brief BMFontConfiguration has parsed configuration of the .fnt file
+@since v0.8
+*/
+class CC_DLL BMFontConfiguration : public Ref
+{
+    // FIXME: Creating a public interface so that the bitmapFontArray[] is accessible
+public://@public
+    // BMFont definitions
+    std::unordered_map<int /* key */, BMFontDef /* fontDef */> _fontDefDictionary;
+
+    //! FNTConfig: Common Height Should be signed (issue #1343)
+    int _commonHeight;
+    //! Padding
+    BMFontPadding    _padding;
+    //! atlas name
+    std::string _atlasName;
+    //! values for kerning
+    std::unordered_map<uint64_t /* key */, int /* amount */> _kerningDictionary;
+
+    // Character Set defines the letters that actually exist in the font
+    std::set<unsigned int>* _characterSet;
+    //! Font Size
+    int _fontSize;
+public:
+    /**
+     * @js ctor
+     */
+    BMFontConfiguration();
+    /**
+     * @js NA
+     * @lua NA
+     */
+    virtual ~BMFontConfiguration();
+    /**
+     * @js NA
+     * @lua NA
+     */
+    std::string description() const;
+
+    /** allocates a BMFontConfiguration with a FNT file */
+    static BMFontConfiguration* create(const std::string& FNTfile);
+
+    /** initializes a BitmapFontConfiguration with a FNT file */
+    bool initWithFNTfile(const std::string& FNTfile);
+
+    const std::string& getAtlasName() { return _atlasName; }
+    void setAtlasName(const std::string& atlasName) { _atlasName = atlasName; }
+
+    std::set<unsigned int>* getCharacterSet() const;
+
+protected:
+    virtual std::set<unsigned int>* parseConfigFile(const std::string& controlFile);
+    virtual std::set<unsigned int>* parseBinaryConfigFile(unsigned char* pData, unsigned long size, const std::string& controlFile);
+
+private:
+    unsigned int parseCharacterDefinition(const char* line);
+    void parseInfoArguments(const char* line);
+    void parseCommonArguments(const char* line);
+    void parseImageFileName(const char* line, const std::string& fntFile);
+    void parseKerningEntry(const char* line);
+    void purgeKerningDictionary();
+    void purgeFontDefDictionary();
+};
 
 class CC_DLL FontFNT : public Font
 {
@@ -47,7 +143,10 @@ public:
     static void purgeCachedData();
     virtual int* getHorizontalKerningForTextUTF32(const std::u32string& text, int &outNumLetters) const override;
     virtual FontAtlas *createFontAtlas() override;
+
     void setFontSize(float fontSize);
+    float getFontSize() const { return _fontSize; }
+
     int getOriginalFontSize()const;
 
     static void reloadBMFontResource(const std::string& fntFilePath);
@@ -60,12 +159,10 @@ protected:
      * @lua NA
      */
     virtual ~FontFNT();
-    
-private:
-    
+
+    BMFontConfiguration* _configuration;
     int  getHorizontalKerningForChars(char32_t firstChar, char32_t secondChar) const;
     
-    BMFontConfiguration * _configuration;
     Vec2                   _imageOffset;
     //User defined font size
     float  _fontSize;
